+++ TopModel.Generator.Jpa/ClassGeneration/JavaEnumEnumGenerator.cs
@@ -1,5 +1,7 @@
 ﻿using Microsoft.Extensions.Logging;
 using TopModel.Core.Model;
+using TopModel.Generator.Core;
+using TopModel.Generator.Jpa.ClassGeneration.Utils;
 using TopModel.Utils;
 
 namespace TopModel.Generator.Jpa.ClassGeneration;
@@ -8,10 +10,10 @@ namespace TopModel.Generator.Jpa.ClassGeneration;
 /// Générateur de fichiers de modèles JPA.
 /// </summary>
 public class JavaEnumEnumGenerator(ILogger<JavaEnumEnumGenerator> logger, IFileWriterProvider writerProvider)
-    : JavaClassGeneratorBase(logger, writerProvider)
+    : ClassGeneratorBase<JpaConfig>(logger, writerProvider)
 {
-    public override string Name => "JavaEnumEnumGen";
-
+    private JavaEnumGeneratorHelper? _javaEnumGeneratorHelper;
+    public override string Name => "JavaDtoGen";
     public override IEnumerable<string> GeneratedFiles =>
         Config
             .Files.Values.SelectMany(f => f.Classes.Where(FilterClass))
@@ -20,140 +22,107 @@ public class JavaEnumEnumGenerator(ILogger<JavaEnumEnumGenerator> logger, IFileW
             )
             .Distinct();
 
+    protected virtual JavaEnumGeneratorHelper JavaEnumGeneratorHelper
+    {
+        get
+        {
+            _javaEnumGeneratorHelper ??= new JavaEnumGeneratorHelper(Config);
+            return _javaEnumGeneratorHelper;
+        }
+    }
+
+    public virtual JavaMethod GetGetter(string tag, IProperty property)
+    {
+        var field = GetField(property, tag);
+        var method = new JavaMethod(field.Type, Config.GetGetterName(property))
+        {
+            Comment = $"Getter for {field.Name}",
+            Body =
+            {
+                new WriterLine() { Line = $"return this.{field.Name};", Indent = 0 },
+            },
+            ReturnComment = $"value of {{@link #{field.Name} {field.Name}}}",
+            Visibility = "public",
+        };
+        method.Imports.AddRange(property.GetTypeImports(Config, tag));
+        return method;
+    }
+
     protected override bool FilterClass(Class classe)
     {
         return !classe.Abstract && classe.Enum == EnumMode.Enum;
     }
 
+    protected virtual IEnumerable<JavaMethod> GetConstuctors(Class classe, string tag)
+    {
+        var constructor = JavaEnumGeneratorHelper.GetConstructor(
+            classe,
+            classe.Properties.Where(p => p.EnumProperty != p),
+            tag
+        );
+        constructor.Visibility = "private";
+        return [constructor];
+    }
+
+    protected virtual IEnumerable<JavaField> GetFields(Class classe, string tag)
+    {
+        foreach (var property in classe.Properties.Where(p => p.EnumProperty != p))
+        {
+            JavaField field = GetField(property, tag);
+            yield return field;
+        }
+    }
+
     protected override string GetFileName(Class classe, string tag)
     {
-        return Config.GetEnumFileName(classe, tag);
+        return Path.Combine(
+            Config.OutputDirectory,
+            Config.ResolveVariables(Config.EnumsPath, tag, module: classe.Namespace.Module).ToFilePath(),
+            $"{classe.NamePascal}.java"
+        );
+    }
+
+    protected virtual IEnumerable<JavaMethod> GetGetters(Class classe, string tag)
+    {
+        if (!(Config.HasAnnotation(classe, "Getter") || Config.HasAnnotation(classe, "Data")))
+        {
+            foreach (var property in Config.GetAvailableProperties(classe).Where(p => p.EnumProperty != p))
+            {
+                if (!Config.HasAnnotation(property, "Getter") || Config.HasAnnotation(classe, "Data"))
+                {
+                    yield return GetGetter(tag, property);
+                }
+            }
+        }
     }
 
     protected override void HandleClass(string fileName, Class classe, string tag)
     {
         var packageName = Config.GetEnumPackageName(classe, tag);
-        using var fw = this.OpenJavaWriter(fileName, packageName, codePage: null);
-        fw.WriteLine();
-        fw.WriteDocStart(0, $"Enumération des valeurs possibles de la classe {classe.NamePascal}");
-        fw.WriteDocEnd(0);
 
-        WriteAnnotations(fw, classe, tag);
-        if (classe.Extends is not null)
+        var javaClass = InitClass(classe, tag);
+
+        using var fw = this.OpenJavaWriter(fileName, packageName, codePage: null);
+        fw.Write(0, javaClass);
+    }
+
+    protected virtual JavaClass InitClass(Class classe, string tag)
+    {
+        var javaEnum = new JavaEnum(classe.NamePascal) { Comment = classe.Comment };
+
+        if (Config.GeneratedHint)
         {
-            fw.AddImport($"{Config.GetPackageName(classe.Extends, tag)}.{classe.Extends.NamePascal}");
+            javaEnum.Add(Config.GeneratedAnnotation);
         }
 
         var implements = Config.GetClassImplements(classe, tag).ToList();
-
-        fw.WriteClassDeclaration(classe.NamePascal, modifier: null, inheritedClass: null, implements, "enum");
-        var i = 0;
-
-        var refs = Config.GetAllValues(classe).ToList();
-
-        var notPkProperties = classe.Properties.Where(p => p != classe.EnumKey);
-        foreach (var refValue in refs)
-        {
-            if (i > 0)
-            {
-                fw.WriteLine();
-            }
-
-            i++;
-            if (classe.DefaultProperty != null)
-            {
-                fw.WriteDocStart(1, $"{refValue.Value[classe.DefaultProperty]}");
-                fw.WriteDocEnd(1);
-            }
-
-            List<string> enumAsString = [];
-            if (!notPkProperties.Any())
-            {
-                enumAsString.Add($"{refValue.Value[classe.EnumKey!]}");
-            }
-            else
-            {
-                enumAsString.Add($"{refValue.Value[classe.EnumKey!]}(");
-                foreach (var prop in notPkProperties)
-                {
-                    var value = refValue.Value.TryGetValue(prop, out var v) ? v : "null";
-                    value = Config.GetValue(prop, value);
-                    if (
-                        prop is { Association: Class association, AssociationProperty: IProperty ap }
-                        && association.Values.Any(r => r.Value.ContainsKey(ap) && r.Value[ap] == value)
-                        && value != "null"
-                    )
-                    {
-                        fw.AddImport(
-                            $"{Config.GetEnumPackageName(association.EnumKey!.Class, tag)}.{association.NamePascal}"
-                        );
-                    }
-
-                    if (
-                        Config.TranslateReferences == true
-                        && classe.DefaultProperty == prop
-                        && prop.EnumProperty == null
-                    )
-                    {
-                        value = @$"""{refValue.ResourceKey}""";
-                    }
-
-                    enumAsString.Add($@"{value}{(prop == notPkProperties.Last() ? string.Empty : ", ")}");
-                }
-
-                enumAsString.Add($")");
-            }
-
-            enumAsString.Add(",");
-
-            fw.WriteLine(1, enumAsString.Aggregate(string.Empty, (acc, curr) => acc + curr));
-        }
-
-        fw.WriteLine();
-        fw.WriteLine(1, ";");
-
-        foreach (var prop in notPkProperties)
-        {
-            fw.WriteLine();
-            fw.WriteDocStart(1, $@"{prop.NamePascal}");
-            fw.WriteDocEnd(1);
-            var fieldName = prop.NameCamel;
-            if (prop is { Association: Class association })
-            {
-                fieldName = $"{prop.NameCamel}";
-                fw.WriteLine(1, $@"private final {association.NamePascal} {fieldName};");
-            }
-            else
-            {
-                fw.WriteLine(1, $@"private final {Config.GetType(prop)} {fieldName};");
-            }
-        }
-
-        if (notPkProperties.Any())
-        {
-            WriteConstructor(classe, fw);
-        }
-
-        foreach (var prop in notPkProperties)
-        {
-            var fieldName = prop.NameCamel;
-            var fieldType = Config.GetType(prop);
-            if (prop is { Association: Class { Enum: EnumMode.Enum } association })
-            {
-                fieldName = $"{prop.NameCamel}";
-                fieldType = $"{association.NamePascal}";
-            }
-
-            var method = new JavaMethod(fieldType, $"get{fieldName.ToFirstUpper()}")
-            {
-                Visibility = "public",
-                Comment = $"Getter for {fieldName}",
-            };
-            method.AddBodyLine($@"return this.{fieldName};");
-            fw.Write(1, method);
-        }
-
-        fw.WriteLine("}");
+        javaEnum.Values.AddRange(GetEnumValues(classe));
+        javaEnum.Implements.AddRange(implements);
+        javaEnum.Imports.AddRange(Config.GetDecoratorImports(classe, tag));
+        javaEnum.AddRange(GetConstuctors(classe, tag));
+        javaEnum.AddRange(GetFields(classe, tag));
+        javaEnum.AddRange(GetGetters(classe, tag));
+        return javaEnum;
     }
 
     private static IEnumerable<IProperty> GetEnumProperties(Class classe)
@@ -161,37 +130,37 @@ public class JavaEnumEnumGenerator(ILogger<JavaEnumEnumGenerator> logger, IFileW
         return classe.Properties.Where(e => e.EnumProperty == e);
     }
 
-    private void WriteAnnotations(JavaWriter fw, Class classe, string tag)
+    private IEnumerable<JavaEnumValue> GetEnumValues(Class classe)
     {
-        fw.AddImports(Config.GetDecoratorImports(classe, tag).ToList());
-        fw.Write(0, GetAnnotations(classe, tag));
+        foreach (var refValue in classe.Values)
+        {
+            var args = classe
+                .Properties.Where(p => p.EnumProperty != p)
+                .Select(prop => JavaEnumGeneratorHelper.GetPropertyValue(classe, prop, refValue))
+                .ToArray();
+            var value = new JavaEnumValue(refValue.Value[classe.EnumKey]);
+            if (
+                classe.DefaultProperty != null
+                && refValue.Value.TryGetValue(classe.DefaultProperty, out var defaultValue)
+            )
+            {
+                value.Comment = defaultValue;
+            }
+            value.Parameters.AddRange(args);
+            yield return value;
+        }
     }
 
-    private void WriteConstructor(Class classe, JavaWriter fw)
+    private JavaField GetField(IProperty property, string tag)
     {
-        // Constructeur
-        var properties = classe.Properties.Where(p => p != classe.EnumKey);
-        var constructor = new JavaConstructor(classe.NamePascal) { Comment = "Enum values constructor" };
-        var methodParams = properties.Select(
-            (prop, index) =>
-            {
-                var fieldName = prop.NameCamel;
-                var fieldType = Config.GetType(prop);
-                if (prop is { Association: Class association })
-                {
-                    fieldName = $"{prop.NameCamel}";
-                    fieldType = $"{association.NamePascal}";
-                }
-
-                return new JavaMethodParameter(fieldType, fieldName) { Final = true };
-            }
-        );
-        constructor.AddParameters(methodParams);
-        foreach (var param in methodParams)
+        var field = new JavaField(
+            Config.GetType(property, forceAssociationPropertyType: Config.UseJdbc),
+            property.NameCamel
+        )
         {
-            constructor.AddBodyLine($@"this.{param.Name} = {param.Name};");
-        }
-
-        fw.Write(1, constructor);
+            Comment = { property.Comment },
+        };
+        field.Imports.AddRange(property.GetTypeImports(Config, tag));
+        return field;
     }
 }