+++ tests/output/jpa/src/main/javagen/restaurant/jpa_identity_feign/enums/restaurant/CategoriePlatCode.java
@@ -12,14 +12,17 @@ public enum CategoriePlatCode {
 	 * Boisson.
 	 */
 	BOISSON,
+
 	/**
 	 * Dessert.
 	 */
 	DESSERT,
+
 	/**
 	 * Entrée.
 	 */
 	ENTREE,
+
 	/**
 	 * Plat principal.
 	 */