+++ tests/output/jpa/src/main/javagen/restaurant/jpa_sequence_server/entities/restaurant/CategoriePlat.java
@@ -39,16 +39,16 @@ import restaurant.jpa_sequence_server.enums.restaurant.CategoriePlatOrdre;
 public class CategoriePlat {
 
 	@Transient
-	public static final CategoriePlat BOISSON = new CategoriePlat(CategoriePlatCode.BOISSON);
+	public static final CategoriePlat BOISSON = new CategoriePlat(CategoriePlatCode.BOISSON, "restaurant.categoriePlat.values.Boisson", CategoriePlatOrdre.Boisson, new BigDecimal(2));
 
 	@Transient
-	public static final CategoriePlat DESSERT = new CategoriePlat(CategoriePlatCode.DESSERT);
+	public static final CategoriePlat DESSERT = new CategoriePlat(CategoriePlatCode.DESSERT, "restaurant.categoriePlat.values.Dessert", CategoriePlatOrdre.Dessert, null);
 
 	@Transient
-	public static final CategoriePlat ENTREE = new CategoriePlat(CategoriePlatCode.ENTREE);
+	public static final CategoriePlat ENTREE = new CategoriePlat(CategoriePlatCode.ENTREE, "restaurant.categoriePlat.values.Entree", CategoriePlatOrdre.Entree, null);
 
 	@Transient
-	public static final CategoriePlat PLAT = new CategoriePlat(CategoriePlatCode.PLAT);
+	public static final CategoriePlat PLAT = new CategoriePlat(CategoriePlatCode.PLAT, "restaurant.categoriePlat.values.Plat", CategoriePlatOrdre.Plat, new BigDecimal(10));
 
 	/**
 	 * Code de la catégorie.
@@ -77,40 +77,17 @@ public class CategoriePlat {
 	private BigDecimal prixMoyen;
 
 	/**
-	 * No arg constructor.
+	 * All args constructor for 'CategoriePlat'.
+	 * @param code Code de la catégorie.
+	 * @param libelle Libellé de la catégorie.
+	 * @param ordre Ordre d'affichage dans le menu.
+	 * @param prixMoyen Prix moyen de la catégorie, à titre indicatif.
 	 */
-	public CategoriePlat() {
-		// No arg constructor
-	}
-
-	/**
-	 * Enum constructor.
-	 * @param code Code dont on veut obtenir l'instance.
-	 */
-	public CategoriePlat(CategoriePlatCode code) {
+	private CategoriePlat(CategoriePlatCode code, String libelle, Integer ordre, BigDecimal prixMoyen) {
 		this.code = code;
-		switch(code) {
-			case BOISSON:
-				this.libelle = "restaurant.categoriePlat.values.Boisson";
-				this.ordre = CategoriePlatOrdre.Boisson;
-				this.prixMoyen = new BigDecimal(2);
-				break;
-			case DESSERT:
-				this.libelle = "restaurant.categoriePlat.values.Dessert";
-				this.ordre = CategoriePlatOrdre.Dessert;
-				this.prixMoyen = null;
-				break;
-			case ENTREE:
-				this.libelle = "restaurant.categoriePlat.values.Entree";
-				this.ordre = CategoriePlatOrdre.Entree;
-				this.prixMoyen = null;
-				break;
-			case PLAT:
-				this.libelle = "restaurant.categoriePlat.values.Plat";
-				this.ordre = CategoriePlatOrdre.Plat;
-				this.prixMoyen = new BigDecimal(10);
-				break;
-		}
+		this.libelle = libelle;
+		this.ordre = ordre;
+		this.prixMoyen = prixMoyen;
 	}
 
 	/**
@@ -147,6 +124,21 @@ public class CategoriePlat {
 	 */
 	public BigDecimal getPrixMoyen() {
 		return this.prixMoyen;
+	}
+
+	/**
+	 * Retourne la valeur de l'énumération pour la clé spécifiée.
+	 * @param code La clé de l'énumération pour laquelle obtenir la valeur.
+	 *
+	 * @return La valeur de l'énumération correspondant à la clé 'Code'.
+	 */
+	public static CategoriePlat getValue(CategoriePlatCode code) {
+		return switch (code) {
+			case BOISSON -> BOISSON;
+			case DESSERT -> DESSERT;
+			case ENTREE -> ENTREE;
+			case PLAT -> PLAT;
+		};
 	}
 
 	/**